requirements.txt
asgiref==3.7.2
Django==5.0.1
django-autoslug==1.9.9
django-dirtyfields==1.9.2
django-reversion==5.0.12
django-phonenumber-field==7.3.0
django-versatileimagefield==3.1
djangorestframework==3.14.0
django-filter==23.5
django-extensions==3.2.3
drf-spectacular==0.27.1
phonenumbers==8.13.29
sqlparse==0.4.4
sorl-thumbnail==12.10.0
